requirements.txt
instaloader
numpy
Pillow
torch
torchvision
Flask
